+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -13,9 +13,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.dto.ReservationTimeRequest;
-import roomescape.controller.dto.ReservationTimeResponse;
-import roomescape.controller.dto.TimeAvailabilityResponse;
+import roomescape.dto.ReservationTimeRequest;
+import roomescape.dto.ReservationTimeResponse;
+import roomescape.dto.TimeAvailabilityResponse;
 import roomescape.service.ReservationAvailabilityService;
 import roomescape.service.ReservationTimeService;
