+++ src/main/java/roomescape/domain/ReservationAvailability.java
@@ -1,4 +1,4 @@
-package roomescape.service.reservation;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
