+++ src/main/java/roomescape/domain/ThemeName.java
@@ -1,4 +1,4 @@
-package roomescape.service.reservation;
+package roomescape.domain;
 
 import java.util.Objects;
