+++ src/main/java/roomescape/dto/ReservationTimeRequest.java
@@ -1,8 +1,8 @@
-package roomescape.controller.dto;
+package roomescape.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
-import roomescape.service.reservation.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 public record ReservationTimeRequest(@NotNull(message = "시간을 입력해주세요.") LocalTime startAt) {
