+++ src/main/java/roomescape/dto/ReservationTimeResponse.java
@@ -1,8 +1,8 @@
-package roomescape.controller.dto;
+package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.service.reservation.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 public record ReservationTimeResponse(Long id, @JsonFormat(pattern = "HH:mm") LocalTime startAt) {
