+++ src/main/java/roomescape/dto/TimeAvailabilityResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
