+++ src/main/java/roomescape/repository/JdbcThemeDao.java
@@ -9,8 +9,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.service.reservation.Theme;
-import roomescape.service.reservation.ThemeName;
+import roomescape.domain.Theme;
+import roomescape.domain.ThemeName;
 
 @Repository
 public class JdbcThemeDao implements ThemeDao {
