+++ src/main/java/roomescape/repository/ReservationTimeDao.java
@@ -3,7 +3,7 @@ package roomescape.repository;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
-import roomescape.service.reservation.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 public interface ReservationTimeDao {
