+++ src/test/java/roomescape/repository/JdbcThemeDaoTest.java
@@ -12,8 +12,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.service.reservation.Theme;
-import roomescape.service.reservation.ThemeName;
+import roomescape.domain.Theme;
+import roomescape.domain.ThemeName;
 
 @JdbcTest
 @Import(JdbcThemeDao.class)
